refactor(dto): restructure topics module and add comment model
