+++ packages/dto/src/topics/mod.rs
@@ -1,0 +1,5 @@
+pub mod comment;
+pub use comment::*;
+
+pub mod topic;
+pub use topic::*;
